pyproject.toml
[project]
name = "qecirc-website"
version = "0.1.0"
description = "QECirc — quantum error correction circuit library"
requires-python = ">=3.13"
dependencies = [
    "numpy",
    "stim",
    "ldpc",
    "mqt-qecc",
    "pyyaml",
    "stimcirq",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]
